% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{build_marginal_frequencies}
\alias{build_marginal_frequencies}
\title{Compute marginal alteration frequencies per sample}
\usage{
build_marginal_frequencies(mt, cnv_wes, del, amp)
}
\arguments{
\item{mt}{Mutation table from \code{\link{read_mutation_table}}.}

\item{cnv_wes}{WES CNV table from \code{\link{read_wes_cnv}}.}

\item{del}{Deletion table from \code{\link{read_wgs_deletions}}.}

\item{amp}{Amplicon table from \code{\link{read_wgs_amplicons}}.}
}
\description{
Compute marginal alteration frequencies per sample
}
