% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{collapse_gi}
\alias{collapse_gi}
\title{Label Pancreas, Stomach, and Colorectal mucinous cancers as GI mucinous}
\usage{
collapse_gi(dat)
}
\description{
Label Pancreas, Stomach, and Colorectal mucinous cancers as GI mucinous
}
