% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{complete_table}
\alias{complete_table}
\title{Create contingency table for comparing differences in mutation rates}
\usage{
complete_table(x, Ns, tumor_order)
}
\description{
Create contingency table for comparing differences in mutation rates
}
