% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{expand_genes}
\alias{expand_genes}
\title{Expand multi-gene entries to one row per gene}
\usage{
expand_genes(x, sep = ", ")
}
\arguments{
\item{x}{Tibble with columns \code{lab_id} and \code{gene_symbol}.}

\item{sep}{Separator used in multi-gene \code{gene_symbol} strings.}
}
\description{
Expand multi-gene entries to one row per gene
}
