% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ggRearrange2}
\alias{ggRearrange2}
\title{Used in 06.1-figure3.rmd to create amplicon graphs}
\usage{
ggRearrange2(df, ylab = "Read pair index", basepairs = 400, num.ticks = 5)
}
\description{
This is a modification to functions with same name (but without '2' postfix)
in svplots package. Instead of returning just a list of grobs, the
modification returns both the grobs and the ggplot objects, allowing further
modification of the ggplot objects prior to creating grobs for the figure.
}
