% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mypca}
\alias{mypca}
\title{Wrapper for principal component analysis of SummarizedExperiment object}
\usage{
mypca(se, scale = FALSE, center = TRUE, rk)
}
\description{
Wrapper for principal component analysis of SummarizedExperiment object
}
