% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{purity_filter}
\alias{purity_filter}
\title{Drop all the samples with purity below threshold}
\usage{
purity_filter(manifest, threshold = 0.2)
}
\description{
Drop all the samples with purity below threshold
}
