% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{read_wes_cnv}
\alias{read_wes_cnv}
\title{Read WES copy-number table (table_S5.tsv) and expand to one row per gene}
\usage{
read_wes_cnv(s5_path)
}
\arguments{
\item{s5_path}{Path to table_S5.tsv.}
}
\description{
Read WES copy-number table (table_S5.tsv) and expand to one row per gene
}
