% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{read_wgs_amplicons}
\alias{read_wgs_amplicons}
\title{Read and process WGS amplicon table (table_s6.rmd/table_s6.tsv)}
\usage{
read_wgs_amplicons(amp_path)
}
\arguments{
\item{amp_path}{Path to table_s6.tsv.}
}
\description{
Read and process WGS amplicon table (table_s6.rmd/table_s6.tsv)
}
