% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{repair_lab_id}
\alias{repair_lab_id}
\title{Make repairs to lab ids}
\usage{
repair_lab_id(tofix, manifest, strip_Ex = FALSE)
}
\arguments{
\item{tofix:}{a single column labeled lab_id}

\item{manifest:}{patient manifest}
}
\value{
a two column tibble with lab_id and prev_id
}
\description{
Make repairs to lab ids
}
