% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{stan_output}
\alias{stan_output}
\title{Stan output}
\usage{
stan_output(
  data.list,
  model,
  params,
  summaryfun,
  probs = c(0.025, 0.05, 0.1, 0.5, 0.9, 0.95, 0.975)
)
}
\description{
Stan output
}
