% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{standardize_pgdx}
\alias{standardize_pgdx}
\title{Standardize pgdx identifiers}
\usage{
standardize_pgdx(dat)
}
\arguments{
\item{dat:}{a tibble with pgdx_id and subject_id field}
}
\description{
Standardize pgdx identifiers
}
