% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{swap_lab_id}
\alias{swap_lab_id}
\title{x contains lab_id
manifest contains lab_id and lab_id2
replace lab_id with lab_id2 when not equal}
\usage{
swap_lab_id(x, y)
}
\description{
x contains lab_id
manifest contains lab_id and lab_id2
replace lab_id with lab_id2 when not equal
}
